requirements.txt
flask
python-dotenv
pinecone-client
transformers
torch
tensorflow
librosa
numpy
scipy
soundfile
